Refactor service build into subprojects
Use modern TypeScript project references and build mode to enable
incremental builds, and make this more similar to the ui project.
