+++ packages/service/src/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../tsconfig.base.json",
+  "compilerOptions": {
+    "outDir": "../dist/src",
+    "tsBuildInfoFile": "src.tsbuildinfo"
+  }
+}
