+++ packages/service/tests/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../tsconfig.base.json",
+  "compilerOptions": {
+    "outDir": "../dist/tests",
+    "tsBuildInfoFile": "tests.tsbuildinfo"
+  }
+}
