+++ packages/service/tests/util/index.ts
@@ -2,7 +2,7 @@ import path from "path";
 import fs from "fs";
 
 export function getTestDataFile(relativePath: string) {
-  return path.join("src/__tests__/__data__/", relativePath);
+  return path.join("tests/data/", relativePath);
 }
 export function readTestDataFile(relativePath: string) {
   return fs.readFileSync(getTestDataFile(relativePath));
