--- packages/service/tsconfig.build.json
@@ -1,9 +1,0 @@
-{
-  "exclude": ["dist/**/*", "scripts/*", "eslint.config.mjs"],
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "noEmit": false,
-    "rootDir": "./src",
-    "outDir": "./dist"
-  }
-}
