+++ packages/ui/.gitignore
@@ -20,6 +20,3 @@ dist-ssr
 *.njsproj
 *.sln
 *.sw?
-
-# TS incremental compilation temp file
-*.tsbuildinfo
